SI-7475 Implement findMembers in terms of findMember
Rather than the copy/paste job that we ended up with after a
performance tuning round leading up to 2.10.

Originally, the delegation was in the other direction:
`findMember` supported use of a wildcard name to grab everything.

While the compiler does spend a fair amount of time in
`findMember` (~5%) ,most of that time seems to be delegated to
`memberType`, `baseClasses`, `lookupEntry`, and `matches`. We didn't
see a measurable improvement from the refactoring to make `findMember`
more specific.

That said, I think the more focussed version is (slightly) easier
to read. So rather than just reverting the change (which turned
out to be tricky for other reasons as callers have grown used to
the slightly different APIs), I've just rewritten `findMembers`
to walk the base class decls and call `findMember` for each new
name it encounters.

`findMembers` isn't hot at all, so this doesn't pose a problem
for performance.

I had to jump through a few hoops to preserve the old order of
members in the returned scope; this is tested widely, albeit
indirectly, in the output of many `neg` tests.