SI-8219 Avoid leaking ambiguous implicit errors from `viewExists`
This method is called during applicability checking, both to
determine if the given arguments are applicable to a candidate,
and to determine which overload to choose.

This was done using the current types Context. The call to
`inferImplicit` disabled immediate reporting of ambiguous implicit
errors, presumably as these should just be treated as a failed
implicit lookup.

But, in some cases, the buffered errors *were* later issued. Worse
still, they were issued at the position of the provided `EmptyTree`,
rather than at the call site being typechecked.

This commit uses silent mode to completely isolate the results of the
implicit search from the caller's Context.