SI-8219 A less roundabout fix for leaking ambigous errors
I took a second look at this after some scrutiny of the preceding
commit, and found that we really want to turn `saveErrors` off
in `viewExists`. That's a lighter weight than silent mode.

That parameter is passed as `saveAmbigousDivergent` to
`inferImplicit`. If that is enabled, any buffered ambiguous or
divergent implicit errors are appended to the outer typer context.
Which is exactly what we *don't* want to happen here!